wip: #205 preparing for remote player position updates
